+++ packages/terra-data-grid/CHANGELOG.md
@@ -5,6 +5,9 @@
   * Added additional screen reader support to announce that column headers are interactable upon selection.
 
 
+* Added
+  * Added ability to resize columns via keyboard.
+
 ## 0.8.1 - (October 17, 2023)
 
 * Fixed